nlp-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.1
